design: Refactor to component
