+++ server/src/components/instgram/MessageInput.tsx
@@ -4,13 +4,19 @@ interface MessageInputProps {
     value: string;
     placeholder: string;
     onChange: (event: React.ChangeEvent<HTMLTextAreaElement>) => void;
+    heightClass?: string; // Tailwind height class
 }
 
-const MessageInput: React.FC<MessageInputProps> = ({ value, placeholder, onChange }) => (
+const MessageInput: React.FC<MessageInputProps> = ({
+    value,
+    placeholder,
+    onChange,
+    heightClass = "h-40",
+}) => (
     <textarea
         value={value}
         onChange={onChange}
-        className="w-full h-40 p-4 rounded-lg resize-none focus:outline-none text-black font-PR_M text-base bg-gray-50 placeholder-gray-300 border-none"
+        className={`w-full p-4 rounded-lg resize-none focus:outline-none text-black font-PR_M text-base bg-gray-50 placeholder-gray-300 border-none ${heightClass}`}
         placeholder={placeholder}
     />
 );
